move test to more-appropriate location and update fixture
